for Factoids, make keys and facts binary datatypes. note that this does not prevent searching.
this is a fix for a bug in all known sqlite using plugins, wherein a \0 gets interpreted as the \x00 byte by python when part of a string,
which is impossible to store in sqlite as part of a text data type. any data cell gets truncated at the first \0 seen.
to fix this, we create a new type "BINARYTEXT", which gets converted to a binary buffer for storage, and subsequently converted back into python string
upon retrieval.
since we had to enable PARSE_DECLTYPES for this, we also use the real types for
factoids.added_at (INTEGER instead of the fake TIMESTAMP) and
factoids.locked (NUMERIC instead of BOOLEAN) (according to section 2.3 of http://www.sqlite.org/datatype3.html)

further, in order to avoid mangling of non-ascii keys in factoids search output, we removed the repr()ing of keys, so that they come through unchanged. which
is actually a fixing of a separate bug, but it came up in the case of the \0s, so we fix it here.

also add testing code for all this. had to pull out some clever tricks to get that done,
since the testing environment is somewhat different from the real irc environment
(see comments in test.py)